refactor: split preview_pane_ui.rs into preview_pane/ modules
