+++ crates/katana-ui/src/lib.rs
@@ -26,7 +26,6 @@ pub use icon::*;
 pub mod changelog;
 pub mod diagram_controller;
 pub mod preview_pane;
-pub mod preview_pane_ui;
 pub mod settings;
 pub mod shell;
 pub mod shell_logic;
