+++ crates/katana-ui/src/preview_pane/mod.rs
@@ -9,6 +9,17 @@ pub mod viewer;
 #[cfg(test)]
 mod tests;
 
+pub(crate) mod utils;
+pub(crate) use utils::*;
+pub(crate) mod section;
+pub(crate) use section::*;
+pub(crate) mod images;
+pub(crate) use images::*;
+pub(crate) mod fullscreen;
+pub(crate) use fullscreen::*;
+pub(crate) mod html;
+pub(crate) mod math;
+pub(crate) use html::*;
 pub use pane::*;
 pub use renderer::*;
 pub use types::*;
