+++ openspec/changes/v0-8-4-rust-code-refactoring/tasks.md
@@ -285,7 +285,7 @@ Tasks Grouped by ## = Adhere unconditionally to the branching standard defined i
 ### 6-D. その他UIファイルの分割
 
 - [x] 6.11 `preview_pane.rs`（1,816行）→ `preview_pane/` サブモジュール
-- [ ] 6.12 `preview_pane_ui.rs`（1,270行）→ `preview/` に統合
+- [x] 6.12 `preview_pane_ui.rs` の分割 (1,270行) -> `preview_pane/` ディレクトリに統合
 - [x] 6.13 `settings_window.rs`（1,666行）→ `settings/` タブごとに分割
 - [ ] 6.14 `i18n.rs`（1,092行）→ `i18n/` サブモジュール
 - [ ] 6.15 `widgets.rs`（948行）→ `widgets/` コンポーネントごとに分割
